+++ src/main/java/dk/i2m/drupal/services/resources/NodeResourceJson.java
@@ -25,7 +25,7 @@ public interface NodeResourceJson {
     @GET("/{alias}/{nid}")
     NodeEntity retrieve(
             @Path("alias") String alias,
-            @Path("nid") Integer nid);
+            @Path("nid") Long nid);
 
     @POST("/{alias}")
     NodeEntity create(
@@ -35,13 +35,13 @@ public interface NodeResourceJson {
     @PUT("/{alias}/{nid}")
     NodeEntity update(
             @Path("alias") String alias,
-            @Path("nid") Integer nid,
+            @Path("nid") Long nid,
             @Body NodeEntity nodeEntity);
 
     @DELETE("/{alias}/{nid}")
     Response delete(
             @Path("alias") String alias,
-            @Path("nid") Integer nid);
+            @Path("nid") Long nid);
 
     @GET("/{alias}")
     List<NodeEntity> index(
@@ -51,14 +51,14 @@ public interface NodeResourceJson {
     @GET("/{alias}/{nid}/files")
     List<FileEntity> files(
             @Path("alias") String alias,
-            @Path("nid") Integer nid);
+            @Path("nid") Long nid);
 
     // TODO: Json equivalent
     @Multipart
     @POST("/{alias}/{nid}/attach_file")
     List<FileEntity> attachFile(
             @Path("alias") String alias,
-            @Path("nid") Integer nid,
+            @Path("nid") Long nid,
             @Part("field_name") String fieldName,
             @Part("attach") Integer attach,
             @PartMap Map<String, Object> params);
